+++ docs/changelog.md
@@ -5,6 +5,7 @@ title: Changelog
 ## 2022-10-19 — New notes
 - **Added**
   - [Git](/software-engineering/git.md) / [Git](/ru/software-engineering/git.md) 🇷🇺
+  - [Web](/software-engineering/Web) / [Web](/ru/software-engineering/Web) 🇷🇺
 
 ## 2022-10-03 — Notes updated
 - **Updated**